fix(config): move process settings shared
Backend layer guards disallow adapters and use_cases importing the\napp layer directly. Move the typed env/process settings owner into\nshared so lower layers can consume it without violating the DAG.\n\nCo-authored-by: Copilot <223556219+Copilot@users.noreply.github.com>
